# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "档案管理系统 API",
        "title": "depositum API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/shelves": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shelf"
                ],
                "summary": "获取架列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.PaginationResponse-github_com_redish101_depositum_pkg_api_v1_Shelf"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shelf"
                ],
                "summary": "创建架",
                "parameters": [
                    {
                        "description": "架",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.CreateShelfRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.Shelf"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/shelves/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shelf"
                ],
                "summary": "获取架详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "架 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.Shelf"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shelf"
                ],
                "summary": "删除架",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "架 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shelf"
                ],
                "summary": "更新架",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "架 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "架",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.UpdateShelfRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.Shelf"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_redish101_depositum_pkg_api_v1.CreateShelfRequest": {
            "type": "object",
            "required": [
                "description",
                "libraryID",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "libraryID": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_redish101_depositum_pkg_api_v1.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_redish101_depositum_pkg_api_v1.LibrarySummary": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_redish101_depositum_pkg_api_v1.PaginationResponse-github_com_redish101_depositum_pkg_api_v1_Shelf": {
            "type": "object",
            "properties": {
                "hasNext": {
                    "description": "是否有下一页",
                    "type": "boolean"
                },
                "hasPrev": {
                    "description": "是否有上一页",
                    "type": "boolean"
                },
                "items": {
                    "description": "数据列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.Shelf"
                    }
                },
                "page": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "每页大小",
                    "type": "integer"
                },
                "total": {
                    "description": "总记录数",
                    "type": "integer"
                },
                "totalPages": {
                    "description": "总页数",
                    "type": "integer"
                }
            }
        },
        "github_com_redish101_depositum_pkg_api_v1.Shelf": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "library": {
                    "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.LibrarySummary"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "github_com_redish101_depositum_pkg_api_v1.UpdateShelfRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "libraryID": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}